app/engine/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
docling
tika
requests
python-docx
pypdf
